chore: extract MarkdownTextWrap from @ourworldindata/grapher
