+++ packages/@ourworldindata/grapher/src/captionedChart/CaptionedChart.tsx
@@ -46,7 +46,7 @@ import { HeaderManager } from "../header/HeaderManager"
 import { SelectionArray } from "../selection/SelectionArray"
 import { EntityName } from "@ourworldindata/core-table"
 import { AxisConfig } from "../axis/AxisConfig"
-import { MarkdownTextWrap, sumTextWrapHeights } from "../text/MarkdownTextWrap"
+import { MarkdownTextWrap, sumTextWrapHeights } from "@ourworldindata/utils"
 
 export interface CaptionedChartManager
     extends ChartManager,
