+++ packages/@ourworldindata/grapher/src/footer/Footer.tsx
@@ -2,12 +2,12 @@ import React from "react"
 import { observable, computed, action } from "mobx"
 import { observer } from "mobx-react"
 import parseUrl from "url-parse"
-import { TextWrap } from "../text/TextWrap"
+import { TextWrap } from "@ourworldindata/utils"
 import { Bounds, DEFAULT_BOUNDS, getRelativeMouse } from "@ourworldindata/utils"
 import { Tooltip } from "../tooltip/Tooltip"
 import { BASE_FONT_SIZE } from "../core/GrapherConstants"
 import { FooterManager } from "./FooterManager"
-import { MarkdownTextWrap } from "../text/MarkdownTextWrap"
+import { MarkdownTextWrap } from "@ourworldindata/utils"
 
 @observer
 export class Footer extends React.Component<{
