+++ packages/@ourworldindata/grapher/src/scatterCharts/ScatterSizeLegend.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { computed } from "mobx"
 import { scaleLinear, ScaleLinear } from "d3-scale"
-import { TextWrap } from "../text/TextWrap"
+import { TextWrap } from "@ourworldindata/utils"
 import { BASE_FONT_SIZE } from "../core/GrapherConstants"
 import { CoreColumn } from "@ourworldindata/core-table"
 import { first, last } from "@ourworldindata/utils"
