+++ packages/@ourworldindata/grapher/src/stackedCharts/MarimekkoChart.tsx
@@ -19,6 +19,7 @@ import {
     SortBy,
     SortConfig,
     SortOrder,
+    TippyIfInteractive,
 } from "@ourworldindata/utils"
 import { action, computed, observable } from "mobx"
 import { observer } from "mobx-react"
@@ -46,7 +47,6 @@ import {
     HorizontalColorLegendManager,
 } from "../horizontalColorLegend/HorizontalColorLegends"
 import { CategoricalBin, ColorScaleBin } from "../color/ColorScaleBin"
-import { TippyIfInteractive } from "../chart/Tippy"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome/index.js"
 import { faInfoCircle } from "@fortawesome/free-solid-svg-icons/faInfoCircle"
 import { DualAxis, HorizontalAxis, VerticalAxis } from "../axis/Axis"
