+++ packages/@ourworldindata/utils/src/DetailsOnDemand/detailsOnDemand.scss
