+++ packages/@ourworldindata/utils/src/MarkdownTextWrap/MarkdownTextWrap.test.ts
@@ -1,6 +1,6 @@
 #! /usr/bin/env jest
 
-import { FontFamily } from "@ourworldindata/utils"
+import { FontFamily } from "../Bounds.js"
 import {
     IRText,
     MarkdownTextWrap,
