+++ site/Footnote.tsx
@@ -1,5 +1,4 @@
-import { Tippy } from "@ourworldindata/grapher"
-import { parseIntOrUndefined } from "@ourworldindata/utils"
+import { parseIntOrUndefined, Tippy } from "@ourworldindata/utils"
 import React from "react"
 import ReactDOM from "react-dom"
