+++ site/covid/index.tsx
@@ -2,8 +2,7 @@ import { faCheckCircle } from "@fortawesome/free-regular-svg-icons/faCheckCircle
 import { faExclamationCircle } from "@fortawesome/free-solid-svg-icons/faExclamationCircle"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome/index.js"
 import { SortOrder } from "@ourworldindata/core-table"
-import { Tippy } from "@ourworldindata/grapher"
-import { oneOf } from "@ourworldindata/utils"
+import { oneOf, Tippy } from "@ourworldindata/utils"
 import classnames from "classnames"
 import React from "react"
 import ReactDOM from "react-dom"
